+++ android/src/main/kotlin/dev/msfjarvis/claw/android/ui/screens/LobstersPostsScreen.kt
@@ -58,7 +58,6 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import androidx.navigation.navDeepLink
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.deliveryhero.whetstone.compose.injectedViewModel
 import com.mikepenz.aboutlibraries.ui.compose.m3.LibrariesContainer
@@ -75,7 +74,6 @@ import dev.msfjarvis.claw.android.ui.navigation.ClawNavigationType
 import dev.msfjarvis.claw.android.ui.navigation.Destinations
 import dev.msfjarvis.claw.android.ui.rememberPostActions
 import dev.msfjarvis.claw.android.viewmodel.ClawViewModel
-import dev.msfjarvis.claw.api.LobstersApi
 import dev.msfjarvis.claw.common.comments.CommentsPage
 import dev.msfjarvis.claw.common.comments.HTMLConverter
 import dev.msfjarvis.claw.common.ui.decorations.ClawAppBar
@@ -228,7 +226,6 @@ fun LobstersPostsScreen(
         enterTransition = { fadeIn(animationSpec = tween(350)) },
         exitTransition = { fadeOut(animationSpec = tween(350)) },
       ) {
-        val uri = LobstersApi.BASE_URL
         composable(route = Destinations.Hottest.route) {
           setWebUri("https://lobste.rs/")
           NetworkPosts(
@@ -252,11 +249,6 @@ fun LobstersPostsScreen(
         composable(
           route = Destinations.Comments.route,
           arguments = listOf(navArgument("postId") { type = NavType.StringType }),
-          deepLinks =
-            listOf(
-              navDeepLink { uriPattern = "$uri/s/${Destinations.Comments.PLACEHOLDER}/.*" },
-              navDeepLink { uriPattern = "$uri/s/${Destinations.Comments.PLACEHOLDER}" },
-            ),
         ) { backStackEntry ->
           val postId =
             requireNotNull(backStackEntry.arguments?.getString("postId")) {
@@ -269,19 +261,31 @@ fun LobstersPostsScreen(
             htmlConverter = htmlConverter,
             getSeenComments = viewModel::getSeenComments,
             markSeenComments = viewModel::markSeenComments,
+            openUserProfile = {
+              navController.navigate(
+                Destinations.User.route.replace(Destinations.User.PLACEHOLDER, it)
+              )
+            },
           )
         }
         composable(
           route = Destinations.User.route,
           arguments = listOf(navArgument("username") { type = NavType.StringType }),
-          deepLinks = listOf(navDeepLink { uriPattern = "$uri/u/${Destinations.User.PLACEHOLDER}" }),
         ) { backStackEntry ->
           val username =
             requireNotNull(backStackEntry.arguments?.getString("username")) {
               "Navigating to ${Destinations.User.route} without necessary 'username' argument"
             }
           setWebUri("https://lobste.rs/u/$username")
-          UserProfile(username = username, getProfile = viewModel::getUserProfile)
+          UserProfile(
+            username = username,
+            getProfile = viewModel::getUserProfile,
+            openUserProfile = {
+              navController.navigate(
+                Destinations.User.route.replace(Destinations.User.PLACEHOLDER, it)
+              )
+            },
+          )
         }
         composable(route = Destinations.Settings.route) {
           SettingsScreen(