+++ common/src/main/kotlin/dev/msfjarvis/claw/common/posts/LobstersCard.kt
@@ -6,6 +6,7 @@
  */
 package dev.msfjarvis.claw.common.posts
 
+import androidx.annotation.VisibleForTesting
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -75,7 +76,7 @@ fun LobstersCard(
           refresh()
         }
         .background(MaterialTheme.colorScheme.background)
-        .padding(start = 16.dp, top = 16.dp, end = 4.dp, bottom = 16.dp)
+        .padding(start = 8.dp)
   ) {
     Row(
       horizontalArrangement = Arrangement.spacedBy(8.dp),
@@ -170,7 +171,7 @@ internal fun Submitter(
 @Composable
 private fun SaveButton(isSaved: Boolean, modifier: Modifier = Modifier) {
   Crossfade(targetState = isSaved, label = "save-button") { saved ->
-    Box(modifier = modifier.padding(12.dp)) {
+    Box(modifier = modifier.padding(vertical = 12.dp)) {
       Icon(
         imageVector = if (saved) Icons.Filled.Favorite else Icons.Filled.FavoriteBorder,
         tint = MaterialTheme.colorScheme.secondary,
@@ -184,7 +185,7 @@ private fun SaveButton(isSaved: Boolean, modifier: Modifier = Modifier) {
 @Composable
 private fun CommentsButton(commentCount: Int?, modifier: Modifier = Modifier) {
   BadgedBox(
-    modifier = modifier.padding(12.dp),
+    modifier = modifier.padding(vertical = 12.dp),
     badge = {
       if (commentCount != null) {
         Badge(
@@ -235,55 +236,55 @@ private fun TagText(tag: String, modifier: Modifier = Modifier) {
   )
 }
 
+@VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+val TEST_POST_ACTIONS =
+  object : PostActions {
+    override fun viewPost(postId: String, postUrl: String, commentsUrl: String) {}
+
+    override fun viewComments(postId: String) {}
+
+    override fun viewCommentsPage(post: UIPost) {}
+
+    override fun toggleSave(post: UIPost) {}
+
+    override suspend fun getComments(postId: String): UIPost {
+      return UIPost(
+        shortId = "ooga",
+        title = "Simple Anomaly Detection Using Plain SQL",
+        url = "https://hakibenita.com/sql-anomaly-detection",
+        createdAt = "2020-09-21T08:04:24.000-05:00",
+        commentCount = 1,
+        commentsUrl = "https://lobste.rs/s/q1hh1g/simple_anomaly_detection_using_plain_sql",
+        tags = listOf("databases", "apis"),
+        description = "",
+        submitter = "Haki",
+        comments = emptyList(),
+      )
+    }
+
+    override suspend fun getLinkMetadata(url: String): LinkMetadata {
+      return LinkMetadata("", "")
+    }
+  }
+
+@VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+val TEST_POST =
+  UIPost(
+    shortId = "ooga",
+    title = "Simple Anomaly Detection Using Plain SQL",
+    url = "https://hakibenita.com/sql-anomaly-detection",
+    createdAt = "2020-09-21T08:04:24.000-05:00",
+    commentCount = 1,
+    commentsUrl = "https://lobste.rs/s/q1hh1g/simple_anomaly_detection_using_plain_sql",
+    submitter = "Haki",
+    tags = listOf("databases", "apis"),
+    description = "",
+    isSaved = true,
+    isRead = true,
+  )
+
 @ThemePreviews
 @Composable
 private fun LobstersCardPreview() {
-  LobstersTheme {
-    LobstersCard(
-      post =
-        UIPost(
-          shortId = "ooga",
-          title = "Simple Anomaly Detection Using Plain SQL",
-          url = "https://hakibenita.com/sql-anomaly-detection",
-          createdAt = "2020-09-21T08:04:24.000-05:00",
-          commentCount = 1,
-          commentsUrl = "https://lobste.rs/s/q1hh1g/simple_anomaly_detection_using_plain_sql",
-          submitter = "Haki",
-          tags = listOf("databases", "apis"),
-          description = "",
-          isSaved = true,
-          isRead = true,
-        ),
-      postActions =
-        object : PostActions {
-          override fun viewPost(postId: String, postUrl: String, commentsUrl: String) {}
-
-          override fun viewComments(postId: String) {}
-
-          override fun viewCommentsPage(post: UIPost) {}
-
-          override fun toggleSave(post: UIPost) {}
-
-          override suspend fun getComments(postId: String): UIPost {
-            return UIPost(
-              shortId = "ooga",
-              title = "Simple Anomaly Detection Using Plain SQL",
-              url = "https://hakibenita.com/sql-anomaly-detection",
-              createdAt = "2020-09-21T08:04:24.000-05:00",
-              commentCount = 1,
-              commentsUrl = "https://lobste.rs/s/q1hh1g/simple_anomaly_detection_using_plain_sql",
-              tags = listOf("databases", "apis"),
-              description = "",
-              submitter = "Haki",
-              comments = emptyList(),
-            )
-          }
-
-          override suspend fun getLinkMetadata(url: String): LinkMetadata {
-            return LinkMetadata("", "")
-          }
-        },
-      refresh = {},
-    )
-  }
+  LobstersTheme { LobstersCard(post = TEST_POST, postActions = TEST_POST_ACTIONS, refresh = {}) }
 }