+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+*   Fixed a crash when clicking an item on the bottom navigation bar
+    too quickly
+
 ## [1.44.0] - 2024-03-19
 
 ### Fixed