+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+*   Upgrade to Compose April releases
+
 ### Fixed
 
 *   Fixed a crash when clicking an item on the bottom navigation bar
     too quickly
+*   Removed buggy deeplinks
+*   Clicking a username now correctly navigates to the right page in-app
 
 ## [1.44.0] - 2024-03-19