+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Remove dividers between posts
+- Post titles in lists are now truncated to be a single line
 
 ### Fixed